Show example MyApp.Guardian.Plug
Unlike `MyApp.Guardian`, there is no explanation of where `MyApp.Guardian.Plug` comes from in `README.md`. I was only able to figure it out by seeing that `Guardian.Plug` had a `__using__` macro and seeing that it accepted a `implementation` argument.  This was a stumbling block for myself and someone I was mentoring, so I've added an example and included the different naming scheme for Phoenix 1.2 vs Phoenix 1.3  I know Guardian really only talks about Plug, but since a lot of developers use it in Phoenix, I thought it was worth tailoring the examples to the differences in the Phoenix versions.  If you don't want to include Phoenix-isms in the docs, I can remove the `MyAppWeb.Guardian.Plug` example.