Allow refreshing tokens without typ field (#670)
The standard "typ" field is not mandatory for JWTs. The token refresh
implementation does not actually rely on the "typ" field in case a
custom TTL is provided, but the `set_ttl` function header mandates the
field.  Instead of matching in the function header, move fetching the
token type into the branch where it is actually necessary.